pyproject.toml
[project]
name = "researchagent"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "openai>=2.3.0",
    "python-dateutil>=2.9.0.post0",
    "python-dotenv>=1.1.1",
    "qwen-agent>=0.0.31",
    "sandbox-fusion>=0.3.7",
]
